Update README with interface functionalities.md
This readme has been updated with descriptions of all the important user interface icons and their functions. The user can interact with the application in various ways, and the readme now details the best way for the user to do so. The readme has listed all the functionalities under the associated image of the icon.